requirements.txt
flask==2.3.3
requests==2.31.0
lxml==4.9.3
flask_cors
rdflib
google-generativeai>=0.7.2